Record SQL Trace durations in microseconds
SqlTracingProxies measured with System.nanoTime() but divided down to
whole milliseconds before recording, so on a local database — where an
ordinary primary-key read finishes in a few hundred microseconds —
almost every execution was stored as 0. Every aggregate built on that
value (stats, statement rankings, percentiles, shares, database time by
request route, the request profile's SQL time) summed truncated zeros,
and the panel correctly reported that it could not rank anything.

Durations are now captured and accumulated in microseconds.
SqlTraceEntryDto carries the exact durationMicros and keeps
durationMillis as a rounded compatibility field; every aggregate
duration field is summed from microseconds and reported in fractional
milliseconds. The slow-query threshold stays configured in milliseconds
and its meaning is unchanged: only the comparison unit moved.

The panel formats durations with a new formatMillis helper that keeps a
sub-millisecond value visible instead of rendering it as 0.

Fixes #1093

Copilot-Session: 75e5f8ad-9a36-47c0-a8ac-8344d9cf77ae

+++ bootui-core/src/main/java/io/github/jdubois/bootui/core/dto/SqlTraceGroupDto.java
@@ -9,8 +9,9 @@ import java.util.List;
  * @param sql the (possibly truncated) statement text shared by the grouped executions
  * @param category coarse SQL category for the statement
  * @param executions number of buffered executions of this exact statement
- * @param totalDurationMillis sum of execution times across the grouped executions
- * @param maxDurationMillis slowest execution time within the group
+ * @param totalDurationMillis sum of execution times across the grouped executions, in fractional
+ *     milliseconds summed from microsecond-resolution executions
+ * @param maxDurationMillis slowest execution time within the group, in fractional milliseconds
  * @param potentialNPlusOne whether the repetition count suggests an N+1 access pattern
  * @param callSites distinct call sites observed for this group's executions, most-recently-seen first
  *     and bounded to a handful of entries; empty when call-site capture is disabled or no application
@@ -20,8 +21,8 @@ public record SqlTraceGroupDto(
         String sql,
         String category,
         long executions,
-        long totalDurationMillis,
-        long maxDurationMillis,
+        double totalDurationMillis,
+        double maxDurationMillis,
         boolean potentialNPlusOne,
         List<String> callSites) {
